Fix rpm build error due to differences in the path
RPM build errors:
  File not found: /home/<user>/rpmbuild/BUILDROOT/cnijfilter-
  ip7200series-3.80-1.x86_64/usr/local/share/cngpijmonip7200/*

  File not found: /home/<user>/rpmbuild/BUILDROOT/cnijfilter-
  ip7200series-3.80-1.x86_64/usr/local/share/maintenanceip7200/*

Here, the target directory is expected to use the printer model
as suffix. However, the template for 'install-data-local' doesn't
use any suffix.

If the variable 'pkgdatadir' was expected to containg desired
suffix, it isn't the case now.

Derive printer model (CNIJ_MODEL) from program-suffix and update
the install template to use it.

New variable was created to prevent any side effects of setting
and/or using existing variable.

Signed-off-by: Sanjeev Premi <spremi@ymail.com>